Added software builds :D (not for dbrebuild)
ForMOM Builder and Runner now have .pyz builds located in the software folder.

For DBRebuild, the user is expected to edit the raw python source,
so building a .pyz makes no sense. In the future it would be good
store configs in .toml files, but for now this is it.